+++ Src/RiderPlugin/src/main/kotlin/com/jetbrains/rider/settings/MismatchedFileNamesOptionsPage.kt
@@ -1,0 +1,7 @@
+package com.jetbrains.rider.settings
+
+import com.jetbrains.rider.settings.simple.SimpleOptionsPage
+
+class MismatchedFileNamesOptionsPage : SimpleOptionsPage(
+        "Mismatched file names",
+        MismatchedFileNamesOptionsPage::class.simpleName!!)