+++ Src/Roflcopter.Plugin/TodoItems/Presentation/TodoItemsCountPresenter.cs
@@ -1,3 +1,4 @@
+#if RESHARPER
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
@@ -98,3 +99,4 @@ namespace Roflcopter.Plugin.TodoItems.Presentation
         }
     }
 }
+#endif