+++ Src/Roflcopter.Plugin/VisualStudio/ModulesWindow/RemoveDebuggingControlIniAction.cs
@@ -1,4 +1,4 @@
-﻿#if !RS20171
+﻿#if !(RS20171 || RIDER)
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;