+++ History.md
@@ -1,3 +1,6 @@
+### vNext ###
+- Added limited support for Rider 2017.2 (see the package description for supported features)
+
 ### 0.11.0 ###
 - Added "Run Unit Tests in File" action
 - Added possibility to write a debugging control .ini file in the Modules Debug window of Visual Studio to disable Jitter optimization of specific modules